Extracted authentication methods to a seperate service class
