+++ backend/src/main/java/com/htecgroup/skynest/service/AuthService.java
@@ -1,0 +1,17 @@
+package com.htecgroup.skynest.service;
+
+import com.htecgroup.skynest.model.dto.UserDto;
+import com.htecgroup.skynest.model.request.UserPasswordResetRequest;
+
+public interface AuthService {
+
+  String confirmEmail(String token);
+
+  String resetPassword(UserPasswordResetRequest userPasswordResetRequest);
+
+  UserDto verifyUser(UserDto userDto);
+
+  void sendVerificationEmail(String emailAddress);
+
+  void sendPasswordResetEmail(String emailAddress);
+}
